Added a Feature
I added a new feature where the letters of the word will be displayed in their correct position in the answer. I still need to work on the guess() function as some functionalities  are broken. I also would like to implement some sort of method to keep track of guesses so the player only has a limited  number of them.